+++ cedar-policy-core/src/pst/err.rs
@@ -36,11 +36,6 @@ use crate::est;
 #[derive(Debug, Clone, PartialEq, Eq, Diagnostic, Error)]
 #[non_exhaustive]
 pub enum PstConstructionError {
-    /// Trying to construct a policy from an empty representation of another type
-    #[error(transparent)]
-    #[diagnostic(transparent)]
-    PolicyFromEmptyRepresentation(#[from] error_body::PolicyFromEmptyRepresentationError),
-
     /// A policy is a linked policy but no link id has been provided
     #[error(transparent)]
     #[diagnostic(transparent)]
@@ -207,11 +202,6 @@ pub mod error_body {
 
     use crate::est;
     use crate::pst;
-
-    /// Trying to construct a policy from an empty representation of another type
-    #[derive(Debug, Clone, PartialEq, Eq, Diagnostic, Error)]
-    #[error("cannot construct policy from empty representation")]
-    pub struct PolicyFromEmptyRepresentationError;
 
     /// A policy is a linked policy but no link id has been provided
     #[derive(Debug, Clone, PartialEq, Eq, Diagnostic, Error)]
@@ -413,6 +403,7 @@ pub mod error_body {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use cool_asserts::assert_matches;
 
     #[test]
     fn from_json_error_conversions() {
@@ -429,7 +420,7 @@ mod tests {
         let json_deser_err: crate::entities::json::err::JsonDeserializationError = serde_err.into();
         let err: PstConstructionError =
             FromJsonError::JsonDeserializationError(json_deser_err).into();
-        assert!(matches!(err, PstConstructionError::ParsingFailed(..)));
+        assert_matches!(err, PstConstructionError::ParsingFailed(..));
 
         // ActionSlot
         let err: PstConstructionError = FromJsonError::ActionSlot.into();
@@ -445,11 +436,11 @@ mod tests {
                 euids: nonempty::nonempty![std::sync::Arc::new(euid)],
             })
             .into();
-        assert!(matches!(err, PstConstructionError::InvalidEntityType(..)));
+        assert_matches!(err, PstConstructionError::InvalidEntityType(..));
 
         // InvalidSlotName
         let err: PstConstructionError = FromJsonError::InvalidSlotName.into();
-        assert!(matches!(err, PstConstructionError::ParsingFailed(..)));
+        assert_matches!(err, PstConstructionError::ParsingFailed(..));
 
         // TemplateToPolicy
         let err: PstConstructionError = FromJsonError::TemplateToPolicy(
@@ -461,7 +452,7 @@ mod tests {
             },
         )
         .into();
-        assert!(matches!(err, PstConstructionError::ContainsSlots(..)));
+        assert_matches!(err, PstConstructionError::ContainsSlots(..));
 
         // PolicyToTemplate
         let err: PstConstructionError = FromJsonError::PolicyToTemplate(
@@ -484,18 +475,18 @@ mod tests {
             },
         )
         .into();
-        assert!(matches!(err, PstConstructionError::ContainsSlots(..)));
+        assert_matches!(err, PstConstructionError::ContainsSlots(..));
 
         // MissingOperator
         let err: PstConstructionError = FromJsonError::MissingOperator.into();
-        assert!(matches!(err, PstConstructionError::InvalidExpression(..)));
+        assert_matches!(err, PstConstructionError::InvalidExpression(..));
 
         // MultipleOperators
         let err: PstConstructionError = FromJsonError::MultipleOperators {
             ops: vec!["a".into(), "b".into()],
         }
         .into();
-        assert!(matches!(err, PstConstructionError::InvalidExpression(..)));
+        assert_matches!(err, PstConstructionError::InvalidExpression(..));
     }
 
     #[test]
@@ -507,7 +498,7 @@ mod tests {
             },
         )
         .into();
-        assert!(matches!(err, PstConstructionError::DuplicateRecordKey(..)));
+        assert_matches!(err, PstConstructionError::DuplicateRecordKey(..));
     }
 
     #[test]
@@ -520,6 +511,6 @@ mod tests {
             },
         )
         .into();
-        assert!(matches!(err, PstConstructionError::UnknownFunction(..)));
+        assert_matches!(err, PstConstructionError::UnknownFunction(..));
     }
 }